Add files via upload
The region in the genome I chose corresponds to Gene LOC119410988 which is an mRNA-eukaryotic translation initiation factor (NCBI). There are 5 RNA seq peaks (pink) that correlate to exons as well. Downstream of these 5 exons is an ATAC-Seq peak (blue) that corresponds to potential regulatory elements of these gene. If you zoom in to the area, you will notice that both the RNA-seq peaks and the ATAC-Seq peaks are higher in the larvae stage than they are in the diapause stage which could mean that this translation initiator is required more frequently in the larvae stage than in the diapause stage.